feat: added function in cron.py for codechef app
